fixed tests, moved ContentFooter to components
