+++ frontend/src/components/ContentFooter.vue
