+++ frontend/src/layouts/DashboardLayout_gdd.vue
@@ -41,7 +41,7 @@
 import Navbar from '@/components/Menu/Navbar.vue'
 import Sidebar from '@/components/Menu/Sidebar.vue'
 import { logout, transactionsQuery } from '@/graphql/queries'
-import ContentFooter from './ContentFooter.vue'
+import ContentFooter from '@/components/ContentFooter.vue'
 import { FadeTransition } from 'vue2-transitions'
 import CONFIG from '@/config'
